+++ gradle.properties
@@ -4,4 +4,3 @@ base_version     = 1.3.0
 mc_version       = 1.8.9
 yarn_version     = 202204162315
 loader_version   = 0.13.3
-testmods_project = testmods
