+++ modules/client/keybindings/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,11 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-fabric-keybinding-api-v1-testmod",
+  "description": "Tests for keybindings",
+  "version": "1.0.0",
+  "entrypoints": {
+    "client": [
+      "net.legacyfabric.fabric.test.client.keybinding.KeybindingTest"
+    ]
+  }
+}
