+++ modules/client/rendering/src/testmod/resources/assets/legacy-fabric-api/lang/en_US.lang
@@ -1,0 +1,3 @@
+# Keybindings
+api.keybinding.testTranslationKey=Test Keybinding
+legacyfabric.api.youreTesting=You're Testing Fabric API!
