+++ modules/client/rendering/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,12 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-fabric-rendering-api-v1-testmod",
+  "description": "Tests for rendering hooks",
+  "version": "1.0.0",
+  "entrypoints": {
+    "client": [
+      "net.legacyfabric.fabric.test.client.rendering.ModelFeatureTest",
+      "net.legacyfabric.fabric.test.client.rendering.RenderingEventsTest"
+    ]
+  }
+}
