+++ modules/entity-events/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,11 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-fabric-entity-events-v1-testmod",
+  "description": "Tests for entity events",
+  "version": "1.0.0",
+  "entrypoints": {
+    "main": [
+      "net.legacyfabric.fabric.test.entity.EntityEventsTest"
+    ]
+  }
+}
