+++ modules/lifecycle-events/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,14 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-fabric-lifecycle-events-v1-testmod",
+  "description": "Tests for lifecycle events",
+  "version": "1.0.0",
+  "entrypoints": {
+    "main": [
+      "net.legacyfabric.fabric.test.lifecycle.ServerLifecycleEventsTest"
+    ],
+    "client": [
+      "net.legacyfabric.fabric.test.lifecycle.client.ClientLifecycleEventsTest"
+    ]
+  }
+}
