+++ modules/resources/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,11 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-fabric-resource-loader-v1-testmod",
+  "description": "Tests for resource loading",
+  "version": "1.0.0",
+  "entrypoints": {
+    "client": [
+      "net.legacyfabric.fabric.test.resource.loader.client.ResourceReloadTest"
+    ]
+  }
+}
