+++ src/main/resources/assets/legacy-fabric-api/icon.png
