--- testmods/src/main/resources/assets/legacy-fabric/blockstates/conc_block.json
@@ -1,7 +1,0 @@
-{
-  "variants": {
-    "normal": {
-      "model": "cake_slice3"
-    }
-  }
-}
