--- testmods/src/main/resources/assets/legacy-fabric/blockstates/conc_block_2.json
@@ -1,7 +1,0 @@
-{
-  "variants": {
-    "normal": {
-      "model": "acacia_leaves"
-    }
-  }
-}
