--- testmods/src/main/resources/assets/legacy-fabric/lang/en_US.lang
@@ -1,5 +1,0 @@
-# Keybindings
-api.keybinding.testTranslationKey=Test Keybinding
-legacyfabric.api.youreTesting=You're Testing Fabric API!
-tile.conc_block.name=Conc Block
-tile.conc_block_2.name=Conc Block 2
