--- testmods/src/main/resources/fabric.mod.json
@@ -1,23 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "legacy-fabric-api-testmods",
-  "description": "Tests for api features",
-  "version": "1.0.0",
-  "entrypoints": {
-    "main": [
-      "net.legacyfabric.fabric.test.ServerLifecycleEventsTest",
-      "net.legacyfabric.fabric.test.ServerNetworkingTest",
-      "net.legacyfabric.fabric.test.EntityEventsTest",
-      "net.legacyfabric.fabric.test.command.SpongeCommandTest",
-      "net.legacyfabric.fabric.test.ConcBlockTest"
-    ],
-    "client": [
-      "net.legacyfabric.fabric.test.client.KeybindingTest",
-      "net.legacyfabric.fabric.test.client.ModelFeatureTest",
-      "net.legacyfabric.fabric.test.client.RenderingEventsTest",
-      "net.legacyfabric.fabric.test.client.ClientNetworkingTest",
-      "net.legacyfabric.fabric.test.client.ClientLifecycleEventsTest",
-      "net.legacyfabric.fabric.test.client.ResourceReloadTest"
-    ]
-  }
-}
